code refactor and tidy up.
