+++ src/main/java/uk/gov/hmcts/sptribs/common/service/AnonymisationService.java
@@ -2,7 +2,18 @@ package uk.gov.hmcts.sptribs.common.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.ccd.sdk.type.FlagDetail;
+import uk.gov.hmcts.ccd.sdk.type.ListValue;
+import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
+import uk.gov.hmcts.sptribs.caseworker.util.CaseFlagsUtil;
+import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
+import uk.gov.hmcts.sptribs.ciccase.model.CicCase;
 import uk.gov.hmcts.sptribs.common.repositories.AnonymisationRepository;
+
+import java.time.LocalDate;
+import java.util.List;
+
+import static uk.gov.hmcts.sptribs.caseworker.util.ErrorConstants.FAILED_TO_ANONYMISE_CASE;
 
 @Service
 @RequiredArgsConstructor
@@ -13,6 +24,47 @@ public class AnonymisationService {
     public String getOrCreateAnonymisation() {
         Long sequence = anonymisationRepository.getNextSequenceValue();
         return sequenceToString(sequence);
+    }
+
+    public void applyAnonymitySelection(CicCase cicCase, List<String> errors, boolean updateAnonymityAlreadyApplied) {
+        if (YesOrNo.YES.equals(cicCase.getAnonymiseYesOrNo())) {
+            if (cicCase.getAnonymisedAppellantName() == null) {
+                String anonymisedName = getOrCreateAnonymisation();
+
+                if (anonymisedName == null) {
+                    errors.add(FAILED_TO_ANONYMISE_CASE);
+                    return;
+                }
+                cicCase.setAnonymisedAppellantName(anonymisedName);
+            }
+            if (cicCase.getAnonymisationDate() == null) {
+                cicCase.setAnonymisationDate(LocalDate.now());
+            }
+            if (updateAnonymityAlreadyApplied) {
+                cicCase.setAnonymityAlreadyApplied(YesOrNo.YES);
+            }
+            return;
+        }
+
+        if (YesOrNo.NO.equals(cicCase.getAnonymiseYesOrNo())) {
+            cicCase.setAnonymisationDate(null);
+            if (updateAnonymityAlreadyApplied) {
+                cicCase.setAnonymityAlreadyApplied(YesOrNo.NO);
+            }
+        }
+    }
+
+    public void processAnonymityFlag(CaseData caseData, CaseData beforeData, List<String> errors) {
+        CicCase cicCase = caseData.getCicCase();
+        if (cicCase == null) {
+            cicCase = new CicCase();
+            caseData.setCicCase(cicCase);
+        }
+
+        ListValue<FlagDetail> effectiveAnonymityFlag = CaseFlagsUtil.mergeAnonymityFlagsPreserveOriginalId(caseData, beforeData);
+        boolean hasActiveAnonymityFlag = CaseFlagsUtil.isActiveFlag(effectiveAnonymityFlag);
+        cicCase.setAnonymiseYesOrNo(hasActiveAnonymityFlag ? YesOrNo.YES : YesOrNo.NO);
+        applyAnonymitySelection(cicCase, errors, true);
     }
 
     /**
